refactor: auto-generate language configuration
